Intern strings created dynamically with concatenation
I had to refactor a bunch of stuff to give the string concatenation code
access to the `Vm`'s internal state (since that's where the set of
interned strings lives).

This is slightly less efficient. Previously we would reuse the buffer
from one `LoxString`'s `String` when concatenating. Now we create a new
buffer. This more closely mirrors the book though.