Manually implement `Hash` for `LoxString`
There's no sense in wasting effort hashing every character in the string
itself when we already did that and stored the result. It'd be nice if I
could somehow just use the stored hash instead of hashing the hash but I
don't think that's possible with the standard library's `Hash` trait.