Fix bug where too many locals were ejected at the end of a scope
We should only pop a local off the stack if its depth is greater than
the `scope_depth` (because that means it was created inside the scope we
just ended (or a child of that scope)). This logic is a little backwards
because we break if it's true. I think in my head I was interpreting it
as part of the `while` logic (where we would break if it was *false*) so
I got the edge case logic wrong.